requirements.txt
# ChatGemini_SakiTool v2.1.2 - 依賴套件清單

# Google Gemini SDK
google-generativeai>=0.3.0  # 傳統 SDK (相容性)
google-genai>=1.45.0         # 新版 SDK (主要使用)

# 環境變數管理
python-dotenv>=1.0.0

# 終端機輸出美化
rich>=13.0.0

# 終端機輸入增強（支援方向鍵、歷史、自動補全）
prompt-toolkit>=3.0.0

# 圖片處理 (用於圖片分析功能)
Pillow>=10.0.0

# 翻譯功能 (思考過程英翻中)
deep-translator>=1.11.4             # 備用翻譯引擎 (免費，無需 API key)
google-cloud-translate>=3.22.0      # 主翻譯引擎 (選用，需要 API key)

# 影片處理 (Veo 功能增強 + Flow Engine)
ffmpeg-python>=0.2.0                # ffmpeg Python wrapper
# opencv-python>=4.8.0              # 場景檢測（選用，未來功能）

# Codebase Embedding (CodeGemini 核心功能)
# 注意：使用自建輕量級向量資料庫（SQLite + NumPy + FAISS），不依賴 ChromaDB
# chromadb>=0.4.0                   # 已棄用：與 Python 3.14 不兼容
# pydantic-settings>=2.0.0          # ChromaDB 依賴項
numpy>=1.24.0                       # 向量計算（必須）
faiss-cpu>=1.7.0                    # 向量索引加速（~30MB，< 50MB 閾值）
                                    # 安裝方式（按平台優先）：
                                    #   macOS:  brew install faiss
                                    #   Linux:  apt/yum/pacman install faiss (視發行版)
                                    #   備用:   pip install faiss-cpu
# 註：SQLite 為 Python 內建，無需額外安裝

# 性能監控 (v1.1.0 功能)
psutil>=5.9.0                       # CPU、記憶體監控

# HTTP 請求（驗證系統需要）
requests>=2.31.0

# 網路搜尋（web_search.py 需要）
duckduckgo-search>=4.0.0

# CodeGemini 依賴（完整版需要，基礎版可選）
pyyaml>=6.0
html2text>=2020.1.16
beautifulsoup4>=4.12.0
cachetools>=5.3.0

# ==========================================
# 性能優化依賴（H1-F7 性能優化模組）
# ==========================================
# 說明：用於啟用異步批次處理，獲得最佳性能（5-96x 提升）
# 記憶體開銷：~5 MB（aiohttp 套件）+ ~0.44 MB（優化模組）
# 智能降級：即使安裝失敗，優化模組也會自動降級到同步版本

# 異步 HTTP 客戶端（啟用完整異步功能）
aiohttp>=3.9.0              # 推薦：功能完整，廣泛使用（~5 MB）
                            # 若安裝失敗會自動降級，不影響核心功能
